# shared/openapi/api-gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CogniBoiler API Gateway",
    "description": "REST and WebSocket gateway of the CogniBoiler digital twin: sign-in and sessions, plant state, PLC commands, simulation control, alarms, history, audit and user administration. Errors are Problem Details (RFC 9457).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/plant": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Get Plant State",
        "description": "The latest plant step: measurements, valves, emissions, health, faults.",
        "operationId": "get_plant_state_api_v1_plant_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantStateResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "Get Simulation Status",
        "operationId": "get_simulation_status_api_v1_simulation_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationStatusResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/scenarios": {
      "get": {
        "tags": [
          "simulation"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_api_v1_simulation_scenarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioListResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/pause": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Pause",
        "operationId": "pause_api_v1_simulation_pause_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/resume": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Resume",
        "operationId": "resume_api_v1_simulation_resume_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/speed": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Set Speed",
        "operationId": "set_speed_api_v1_simulation_speed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/step": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Step",
        "description": "Advance a paused simulation by whole steps.",
        "operationId": "step_api_v1_simulation_step_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StepRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/scenario": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Load Scenario",
        "description": "Restart the plant from a scenario; the PLC follows the new run.",
        "operationId": "load_scenario_api_v1_simulation_scenario_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/faults": {
      "post": {
        "tags": [
          "simulation"
        ],
        "summary": "Inject Fault",
        "description": "Inject a labelled fault; the physics engine validates target and severity.",
        "operationId": "inject_fault_api_v1_simulation_faults_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FaultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "simulation"
        ],
        "summary": "Clear All Faults",
        "operationId": "clear_all_faults_api_v1_simulation_faults_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/simulation/faults/{fault_id}": {
      "delete": {
        "tags": [
          "simulation"
        ],
        "summary": "Clear Fault",
        "operationId": "clear_fault_api_v1_simulation_faults__fault_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "fault_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Fault Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaultAckResponse"
                }
              }
            }
          },
          "503": {
            "description": "upstream.unavailable: the upstream service (named in `service`) failed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "upstream.timeout: the upstream service did not answer in time; a command may still have been carried out.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlantStateResponse": {
        "properties": {
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms"
          },
          "simulation": {
            "$ref": "#/components/schemas/SimulationStatusResponse"
          },
          "boiler": {
            "$ref": "#/components/schemas/BoilerStateResponse"
          },
          "turbine": {
            "$ref": "#/components/schemas/TurbineStateResponse"
          },
          "actuators": {
            "$ref": "#/components/schemas/ActuatorStateResponse"
          },
          "emissions": {
            "$ref": "#/components/schemas/EmissionsResponse"
          },
          "condenser": {
            "$ref": "#/components/schemas/CondenserResponse"
          },
          "health": {
            "$ref": "#/components/schemas/EquipmentHealthResponse"
          },
          "performance": {
            "$ref": "#/components/schemas/PerformanceResponse"
          },
          "faults": {
            "items": {
              "$ref": "#/components/schemas/FaultResponse"
            },
            "type": "array",
            "title": "Faults"
          },
          "sensors": {
            "items": {
              "$ref": "#/components/schemas/SensorStatusResponse"
            },
            "type": "array",
            "title": "Sensors"
          }
        },
        "type": "object",
        "required": [
          "timestamp_ms",
          "simulation",
          "boiler",
          "turbine",
          "actuators",
          "emissions",
          "condenser",
          "health",
          "performance",
          "faults",
          "sensors"
        ],
        "title": "PlantStateResponse",
        "description": "Everything the physics engine publishes about the plant at one step."
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Stable, machine-readable error code."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "ProblemDetails",
        "description": "An error body (RFC 9457); members beyond these depend on the code."
      },
...
      "SpeedRequest": {
        "properties": {
          "speed_factor": {
            "type": "number",
            "maximum": 50.0,
            "exclusiveMinimum": 0.0,
            "title": "Speed Factor",
            "description": "Simulated seconds per wall-clock second."
          }
        },
        "type": "object",
        "required": [
          "speed_factor"
        ],
        "title": "SpeedRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}